Safari-609.1.11.1/LayoutTests/js/arrowfunction-block-1-expected.txt

Tests for ES6 arrow function declaration body as block

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


f = () => {}
PASS typeof f() is 'undefined'
g = () => ({})
PASS typeof g() is 'object'
PASS successfullyParsed is true

TEST COMPLETE
